Launch standalone YouTube auth server and wire PS1 MDEC
What changed:
- Landed the new Node/Express YouTube auth proxy with docs, env ignores, logging defaults, and README wiring so Google secrets stay outside the Qt app.
- Rebuilt the big-screen YouTube experience (QSS, browse/player overlays, streaming web views) with device-login rails, sidebar navigation, QR auth, and playback metadata pulled from the new service.
- Upgraded the PS1 stack with an MDEC implementation, richer GPU/SPU diagnostics, pad-buffer tracking, and DMA/CDROM hooks so FMV playback and controller polling match hardware expectations.
- Modernized the CMake/Ninja toolchain: ccache launcher, sentinel-driven configure, Qt Multimedia linkage, precompiled headers, and macOS signing entitlements plus new VideoRecorder accessors.

Why this matters:
- Secrets finally live in a managed server, letting us rotate credentials and unblock OAuth scopes without touching the desktop build.
- The TV UI now feels like an actual streaming box—account-aware guide, auth flows, recommendations, and a media player you can drive with a d-pad.
- PS1 titles that depend on MDEC video and precise pad DMA (Crash FMVs, BIOS PAD drivers) now boot and run with far fewer hacks.
- Faster iteration and deterministic builds keep developer productivity up even as the project grows.

Breaking changes:
- Likely breaking change detected; review staged diff

Risk: high (breaking API changes detected)

Testing: fail (degraded)

+++ include/emulator/ps1/GTE.h
@@ -43,11 +43,11 @@ private:
   int16_t sxy[4][2] = {}; // SXY0-3 FIFO (X,Y signed 16-bit)
   uint16_t sz[4] = {};    // SZ0-3 FIFO (unsigned 16-bit)
   uint32_t rgb[3] = {};   // RGB0-2 FIFO
-  int32_t mac[4] = {};    // MAC0-3 (32-bit, but internal calcs use 44-bit)
-  uint32_t irgb = 0;      // Color conversion input
-  uint32_t orgb = 0;      // Color conversion output
-  int32_t lzcs = 0;       // Leading zero count source
-  int32_t lzcr = 0;       // Leading zero count result
+  int64_t mac[4] = {}; // MAC0-3 (MAC1-3 keep full internal accumulator width)
+  uint32_t irgb = 0;   // Color conversion input
+  uint32_t orgb = 0;   // Color conversion output
+  int32_t lzcs = 0;    // Leading zero count source
+  int32_t lzcr = 0;    // Leading zero count result
 
   // ─── Control Registers (32 × 32-bit) ────────────────────────────────
   int16_t rt[3][3] = {}; // Rotation matrix 3×3
